Refactor: Consolidate LLM models in schema.py and fix CLI command alias support\n\n- Move and unify LLMConfig and LLMResponse in schema.py\n- Remove duplicates from llm/base.py, import from schema\n- Register CLI command aliases and update tab completion to include them\n- Fixes alias usage for help, exit, load-agent, and more
