+++ spoon_ai/schema.py
@@ -53,8 +53,21 @@ class Message(BaseModel):
 TOOL_CHOICE_VALUES = tuple(choice.value for choice in ToolChoice)
 TOOL_CHOICE_TYPE = Literal[TOOL_CHOICE_VALUES] # type: ignore
 
+class LLMConfig(BaseModel):
+    """Configuration for LLM providers"""
+    model: str = ""
+    api_key: str = ""
+    base_url: Optional[str] = None
+    api_type: Optional[str] = None
+    api_version: Optional[str] = None
+    max_tokens: int = 4096
+    temperature: float = 0.3
+
 class LLMResponse(BaseModel):
+    """Unified LLM response model"""
     content: str
+    text: str = ""  # Original text response
     tool_calls: List[Any] = Field(default_factory=list)
+    image_paths: List[dict] = Field(default_factory=list)
     finish_reason: Optional[str] = Field(default=None)
     native_finish_reason: Optional[str] = Field(default=None)
